requirements.txt
requests
beautifulsoup4
pyyaml
lxml
nltk